Create first-web-prospect-soup-truck
This is my first project at SheCodes Workshop Basics. I did not have any previous experience in coding. It's not perfect, but I think in the timetable given (1 week) that was what I could come out with. 

My idea was to solve a problem that many restaurant businesses are experiencing right after they were allowed to reopen during the COVID-19 outbreak. I saw many struggling to find people to work. I thought that if they have a website/app with menu options that people could choose even from outside of the restaurant, it would solve the immediate labor shortage. 

It's simple. There are some options. The color change when the person hovers the mouse. Once click on the button, javascript will run asking the name and the option. After javascript change the div element and insert some HTML content. 

I know there is still a lot of room to improve, so please, if you have any suggestions let me know in the comments. I would love to receive some feedback as well. 

Thanks, 
Thais